requirements/base.txt
# Core framework
fastapi>=0.135.0
uvicorn>=0.41.0
pydantic>=2.12.0
pydantic-settings>=2.13.0

# Database
sqlalchemy>=2.0.48
psycopg2-binary>=2.9.11
alembic>=1.14.0

# Auth
python-jose>=3.5.0
bcrypt>=4.0.1
python-multipart>=0.0.22

# LLM — Groq (cloud)
langchain>=1.2.0
langchain-groq>=1.1.0
langchain-community>=0.4.0
langchain-core>=1.2.0
python-dotenv>=1.2.0

# LLM — LangGraph (agent orchestration)
langgraph>=1.0.10
langgraph-checkpoint>=4.0.0
langgraph-prebuilt>=1.0.8

# Search
duckduckgo-search>=8.1.0

# MCP
fastmcp>=3.1.0
mcp>=1.26.0

# Cache / message broker
redis>=7.2.0

# ML — drift detection
numpy>=2.4.0
scikit-learn>=1.8.0
scipy>=1.17.0
